finish 2

+++ app/Http/Controllers/DocentesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+// use Request;
 use App\Docente;
 use App\Grupo;
 use App\Persona;
@@ -11,6 +12,7 @@ use Illuminate\Support\Facades\DB;
 use App\User;
 use App\Http\Requests\ValidarCrearDocenteRequest;
 use App\Http\Controllers\Auth;
+use Barryvdh\DomPDF\Facade as PDF;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Foundation\Auth\AuthenticatesUsers;
@@ -83,6 +85,7 @@ class DocentesController extends Controller
     public function store(ValidarCrearDocenteRequest $request)
     {
         $data = request()->all(); 
+    //    dd($data);
         $agregar = Persona::firstOrCreate([
             'curp' => $data['curp']
         ],[
@@ -103,15 +106,44 @@ class DocentesController extends Controller
         $tipo->tipo = $data['tipo'];
         $tipo->save();
             
-        Docente::firstOrCreate([
+        $docente = Docente::firstOrCreate([
             'curp_docente' => $data['curp']
         ],[
-            'rfc' => $data['rfc'],
+            'rfc' => $data['rfc']->getClientOriginalName(),
             'grado_estudios' => $data['estudios'],
-            'titulo' => $data['titulo'],
-            'ced_prof' => $data['cedula'],
+            'titulo' => $data['titulo']->getClientOriginalName(),
+            'ced_prof' => $data['cedula']->getClientOriginalName(),
+            // 'certificaciones' => $certificaciones,
             'estatus' => $data['estatus']
         ]);
+        $fileRFC = $docente->id_docente.$data['rfc']->getClientOriginalName();
+        $data['rfc']->storeAs('documentosCLE',$fileRFC);
+
+        $fileCedula = $docente->id_docente.$data['cedula']->getClientOriginalName();
+        $data['cedula']->storeAs('documentosCLE',$fileCedula);
+
+        $fileTitulo = $docente->id_docente.$data['titulo']->getClientOriginalName();
+        $data['titulo']->storeAs('documentosCLE',$fileTitulo);
+
+        if ($request->hasFile('certificaciones')) {
+            $certificaciones ="";
+                foreach ($data['certificaciones'] as $cert ) {
+                    $certificaciones = $certificaciones.','.$cert->getClientOriginalName();
+                    $fileName = $docente->id_docente.$cert->getClientOriginalName();
+                    $cert->storeAs('documentosCLE',$fileName);
+                    
+                }
+                $certificaciones = substr($certificaciones,1);
+                $docente->certificaciones = $certificaciones;
+                $docente->save();
+
+                // $documentos = split(',',$certificaciones);
+                // foreach ($documentos as $docs) {
+                //     $fileName = time().$docs;
+                    
+                // }   
+            
+        }
 
         User::firstOrCreate([
             'curp_user' => $data['curp']
@@ -162,8 +194,13 @@ class DocentesController extends Controller
                         ->get();
                        
         $nombres_municipios = Municipio::select('nombre_municipio')->pluck('nombre_municipio');
-
         $email = User::where('curp_user',$docente[0]->curp)->get();
+        if ($email->isNotEmpty()) {
+            $email = $email[0]->email;
+        }else{
+            $email = " ";
+        }
+        
         // dd($docente);
         return view('docentes.editDocente',compact('docente','nombres_municipios','email','usuarioactual'));
     }
@@ -176,8 +213,8 @@ class DocentesController extends Controller
         //obtener los datos de la misma persona en todas las tablas relacionadas
         $persona = Persona::where('curp',$curp)->first();
         $infoDocente= Docente::find($docente->id_docente);
-        $user = User::where('curp_user',$curp)->first();    
-        // dd($infoDocente);
+        $user = User::where('curp_user',$curp)->first();  
+        // dd($persona,$infoDocente,$user);
         //validacion de los datos
         if ($usuarioactual->tipo == 'docente') {
             $data = request()->validate([
@@ -190,10 +227,17 @@ class DocentesController extends Controller
                 'colonia' => 'required|alpha_spaces',
                 'municipio' => 'required',
                 'cp' => 'required|numeric',
-                'telefono' =>'required|numeric',
-                'email' => array('required','email','regex:/^[A-z0-9\\._-]+@[A-z0-9][A-z0-9-]*(\\.[A-z0-9_-]+)*\\.([A-z]{2,6})$/'),
+                'telefono' =>'sometimes|nullable|numeric',
+                'certificaciones' => 'sometimes|nullable',
+                'estudios' => 'sometimes|nullable',
+                'email' => array('sometimes','nullable','email','regex:/^[A-z0-9\\._-]+@[A-z0-9][A-z0-9-]*(\\.[A-z0-9_-]+)*\\.([A-z]{2,6})$/'),
                 'edad' =>'required|digits:2',
-                'sexo' => 'required'
+                'sexo' => 'required',
+                'dominio' => 'sometimes|nullable',
+                'curso' => 'sometimes|nullable',
+                'didactica' => 'sometimes|nullable',
+                'experiencia' => 'sometimes|nullable',
+                'actualizacion' => 'sometimes|nullable'
             ]);
         }else{
             $data = request()->validate([
@@ -206,25 +250,92 @@ class DocentesController extends Controller
                 'colonia' => 'required|alpha_spaces',
                 'municipio' => 'required',
                 'cp' => 'required|numeric',
-                'telefono' =>'required|numeric',
-                'email' => array('required','email','regex:/^[A-z0-9\\._-]+@[A-z0-9][A-z0-9-]*(\\.[A-z0-9_-]+)*\\.([A-z]{2,6})$/'),
+                'telefono' =>'sometimes|nullable|numeric',
+                'email' => array('sometimes','nullable','email','regex:/^[A-z0-9\\._-]+@[A-z0-9][A-z0-9-]*(\\.[A-z0-9_-]+)*\\.([A-z]{2,6})$/'),
                 'edad' =>'required|digits:2',
                 'sexo' => 'required',
-                'rfc' => 'required',
-                'estudios' => 'required',
-                'titulo' => 'required',
-                'cedula' => 'required',
-                'estatus' => 'required'
+                'rfc' => 'sometimes|nullable',
+                'certificaciones' => 'sometimes|nullable',
+                'estudios' => 'sometimes|nullable',
+                'titulo' => 'sometimes|nullable',
+                'cedula' => 'sometimes|nullable',
+                'estatus' => 'required',
+                'dominio' => 'sometimes|nullable',
+                'curso' => 'sometimes|nullable',
+                'didactica' => 'sometimes|nullable',
+                'experiencia' => 'sometimes|nullable|numeric',
+                'actualizacion' => 'sometimes|nullable'
             ]);
-            $infoDocente->rfc = $data['rfc'];
-            $infoDocente->grado_estudios = $data['estudios'];
-            $infoDocente->ced_prof = $data['cedula'];
-            $infoDocente->titulo = $data['titulo'];
-            $infoDocente->estatus = $data['estatus'];
-            $infoDocente->save();
+
+            $c = request()->all();
+            // dd($data,$c);
+            if(request()->has('certificaciones') && $c['certificaciones'][0]!=null){
+                $certificaciones ="";
+                foreach ($c['certificaciones'] as $cert ) {
+                    $certificaciones = $certificaciones.','.$cert->getClientOriginalName();
+                    $fileCertificaciones = $cert;    
+                    $fileName = $docente->id_docente.$fileCertificaciones->getClientOriginalName();
+                    
+                    $fileCertificaciones->move(storage_path('app').'/documentosCLE',$fileName);
+                    // $fileCertificaciones->move(public_path().'/documentosCLE/',$fileName);
+                }
+                $certificaciones = substr($certificaciones,1);
+                $infoDocente->certificaciones = $certificaciones;
+            }
         }
+            
+        if (request()->has('dominio')) {
+            $infoDocente->dominio_idioma = $data['dominio'];
+        }
+        if (request()->has('curso')) {
+            $infoDocente->curso = $data['curso'];
+        }
+        if (request()->has('didactica') && $data['didactica'][0] != null) {
+            $datos_didacticas = "";
+            foreach ($data['didactica'] as $didactica ) {
+                $datos_didacticas = $datos_didacticas.';'.$didactica;
+            }
+            $datos_didacticas = substr($datos_didacticas, 1);
+            // return $datos_didacticas;
+            $infoDocente->didactica = $datos_didacticas;
+        }
+        if (request()->has('experiencia')) {
+            $infoDocente->experiencia = $data['experiencia'];
+        }
+        if (request()->has('actualizacion') && $data['actualizacion'][0] != null) {
+            $actualizaciones = "";
+            foreach ($data['actualizacion'] as $actualizacion ) {
+                $actualizaciones = $actualizaciones.';'.$actualizacion;
+            }
+            $actualizaciones = substr($actualizaciones, 1);
+            // return $actualizaciones;
+            $infoDocente->actualizacion = $actualizaciones;
+        }
+        // return 'fin';
+            if (request()->has('cedula')) {
+                $infoDocente->ced_prof = $data['cedula']->getClientOriginalName();
+                $fileCedula = $data['cedula'];    
+                $fileName = $docente->id_docente.$fileCedula->getClientOriginalName();
+                $fileCedula->move(storage_path('app').'/documentosCLE',$fileName);
+            }
+            if (request()->has('titulo')) {
+                $infoDocente->titulo = $data['titulo']->getClientOriginalName();
+                $fileTitulo = $data['titulo'];    
+                $fileName = $docente->id_docente.$fileTitulo->getClientOriginalName();
+                $fileTitulo->move(storage_path('app').'/documentosCLE',$fileName);
+            }
+            if (request()->has('rfc')) {
+                $infoDocente->rfc = $data['rfc']->getClientOriginalName();
+                $fileRFC = $data['rfc'];    
+                $fileName = $docente->id_docente.$fileRFC->getClientOriginalName();
+                $fileRFC->move(storage_path('app').'/documentosCLE',$fileName);
+            }
+            $infoDocente->grado_estudios = $data['estudios'];
+            // $infoDocente->estatus = $data['estatus'];
+            $infoDocente->save();
+        
         // dd(request()->apMaterno);
-        // $persona->curp = $data['curp'];
+        $persona->curp = $data['curp'];
         $persona->nombres = $data['name'];
         $persona->ap_paterno = $data['apPaterno'];
         $persona->ap_materno = request()->apMaterno;
@@ -264,7 +375,7 @@ class DocentesController extends Controller
 
         DB::update('update personas set curp = ? where curp = ?', [$data['curp'],$curp]);
 
-        if($usuarioactual->tipo = 'docente'){
+        if($usuarioactual->tipo == 'docente'){
             return back()->with('success','¡Tus datos se actualizaron correctamente!');
         }else{
             return redirect()->route('verDocentes')->with('success','¡Los datos se actualizaron correctamente!');
@@ -278,32 +389,49 @@ class DocentesController extends Controller
         $docente = Docente::where('id_docente',$request->docente_id)->get();
 
         $persona = Persona::where('curp',$docente[0]->curp_docente)->get();
-// dd($docente,$persona);
+
         $user = User::where('curp_user',$docente[0]->curp_docente)->get();
-        $user[0]->delete();
-        $docente[0]->delete();
-        $persona[0]->delete();
+        // dd($docente,$persona,$user);
+        if($user->isNotEmpty()){
+            $user[0]->delete();
+        }
+        if($docente->isNotEmpty()){
+            $docente[0]->delete();
+        }
+        if($persona->isNotEmpty()){
+            $persona[0]->delete();
+        }
+        
         // $clave= $docente->curp_docente;
         // $persona = Persona::where('curp','=', $clave);
         // $persona->delete();
         return redirect()->route('verDocentes')->with('warning','Los datos se eliminaron');
     }
 
-    public function titulo($titulo){
-        //se cambiara la ruta de los archivos
-        $p = 'C:\Users\Mony\Downloads/'.$titulo;
+    public function titulo($id,$titulo){
+        $p = storage_path('app').'/documentosCLE/'.$id.$titulo;
         return response()->file($p);
     }
 
-    public function rfc($rfc){
-        //se cambiara la ruta de los archivos
-        $p = 'C:\Users\Mony\Downloads/'.$rfc;
+    public function rfc($id,$rfc){
+        $p = storage_path('app').'/documentosCLE/'.$id.$rfc;
         return response()->file($p);
     }
 
-    public function cedula($cedula){
-        //se cambiara la ruta de los archivos
-        $p = 'C:\Users\Mony\Downloads/'.$cedula;
+    public function cedula($id,$cedula){
+        $p = storage_path('app').'/documentosCLE/'.$id.$cedula;
         return response()->file($p);
+    }
+
+    public function certificaciones($id,$certificacion){
+        $p = storage_path('app').'/documentosCLE/'.$id.$certificacion;
+        return response()->file($p);
+    }
+
+    public function plantilla(){
+        $docentes = Docente::where('estatus','Activo')->leftjoin('personas','personas.curp','=','docentes.curp_docente')->get();
+        // return view('pdf.plantillapdf',compact('docentes')); setPaper('A4','landscape')->
+        $pdfPlantilla =  PDF::loadView('pdf.plantillapdf',compact('docentes'));
+        return $pdfPlantilla->download('PlantillaDocentes-'.strftime("%b%Y").'.pdf');
     }
 }
